requirements.txt
biopython==1.81
pandas==2.2.2
